Rename `text_ui` to `data_label_for_ui_layout` and used it in many places instead of `label_for_ui_layout`
